fix host spray race condition (#3213)
* core: bug fixes

* best practices: uniform comments
